third_party/blink/web_tests/wpt_internal/prerender/postmessage-to-service-worker.https.html

<!DOCTYPE html>
<title>postMessage() between service worker and prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>
const uid = token();

const PAGE_URL = `resources/postmessage-to-service-worker.html?uid=${uid}`;
const WORKER_URL = 'resources/postmessage-to-client-worker.js';

// Message sequence: prerendered page => service worker => prerendered page =>
// main page
promise_test(async t => {
  const registration =
    await service_worker_unregister_and_register(t, WORKER_URL, PAGE_URL);
  t.add_cleanup(() => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');

  const bc = new PrerenderChannel('prerender-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  startPrerendering(PAGE_URL);

  const result = await gotMessage;
  assert_equals(result, 'postmessage to client');
}, 'postMessage() should succeed between a service worker and a prerendered ' +
   'page');

</script>
</body>
